style: renaming and good hrserver start
