+++ examples/main.py
@@ -1,6 +1,6 @@
 import os,sys; sys.path.insert(0,os.path.realpath(os.path.dirname(os.path.dirname(__file__))))
 
-from htagweb import AppServer
+from htagweb import SimpleServer
 from starlette.responses import HTMLResponse
 
 import app1
@@ -43,7 +43,7 @@ async def handlePath(request):
         return HTMLResponse("404",404)
 
 
-app=AppServer()
+app=SimpleServer()
 app.add_route("/{path:path}", handlePath )
 
 if __name__=="__main__":
