+++ htagweb/__init__.py
@@ -7,8 +7,8 @@
 # https://github.com/manatlan/htagweb
 # #############################################################################
 
-from .redysserver import RedysServer  #NOT THE DEFINITIVE NAME !!!!!!!!!!!!!!!!
-from .appserver import AppServer  # a completly different beast, but compatible with ^^
+from .appserver import AppServer
+from .simpleserver import SimpleServer  # a completly different beast, but compatible with ^^
 from .htagserver import HtagServer  # a completly different beast.
 from .usot import Usot
