+++ htagweb/appserver.py
@@ -9,37 +9,15 @@
 
 # gunicorn -w 4 -k uvicorn.workers.UvicornWorker -b localhost:8000 --preload basic:app
 
-"""
-IDEM que htagweb.AppServer
-- mais sans le SHARED MEMORY DICT (donc compat py3.7) ... grace au sesprovider !
-- fichier solo
-- !!! utilise crypto de htagweb !!!
-
-
-
-This thing is completly new different beast, and doesn't work as all classic runners.
-
-It's a runner between "WebHTTP/WebServer" & "HtagServer" : the best of two worlds
-
-    - It's fully compatible with WebHTTP/WebServer (provide a serve method)
-    - it use same techs as HTagServer (WS only, parano mode, simple/#workers, etc ...)
-    - and the SEO trouble is faked by a pre-fake-rendering (it create a hr on http, for seo ... and recreate a real one at WS connect)
-
-Like HTagServer, as lifespan of htag instances is completly changed :
-htag instances should base their state on "self.root.state" only!
-Because a F5 will always destroy/recreate the instance.
-"""
 
 import os
 import sys
 import json
 import uuid
-import pickle
-import inspect
 import logging
 import uvicorn
-import importlib
-import contextlib
+import asyncio
+import multiprocessing
 from htag import Tag
 from starlette.applications import Starlette
 from starlette.responses import HTMLResponse
@@ -51,9 +29,11 @@ from starlette.requests import HTTPConnection
 from starlette.datastructures import MutableHeaders
 from starlette.types import ASGIApp, Message, Receive, Scope, Send
 
-from htag.render import HRenderer
 from htag.runners import commons
 from . import crypto,usot
+
+from htagweb.server import importClassFromFqn, hrserver, hrserver_orchestrator
+from htagweb.server.client import HrPilot
 
 logger = logging.getLogger(__name__)
 ####################################################
@@ -80,31 +60,7 @@ def findfqn(x) -> str:
 
     return tagClass.__module__+"."+tagClass.__qualname__
 
-def getClass(fqn_norm:str) -> type:
-    assert ":" in fqn_norm
-    #--------------------------- fqn -> module, name
-    modulename,name = fqn_norm.split(":",1)
-    if modulename in sys.modules:
-        module=sys.modules[modulename]
-        try:
-            module=importlib.reload( module )
-        except ModuleNotFoundError:
-            """ can't be (really) reloaded if the component is in the
-            same module as the instance htag server"""
-            pass
-    else:
-        module=importlib.import_module(modulename)
-    #---------------------------
-    klass= getattr(module,name)
-    if not ( inspect.isclass(klass) and issubclass(klass,Tag) ):
-        raise Exception(f"'{fqn_norm}' is not a htag.Tag subclass")
 
-    if not hasattr(klass,"imports"):
-        # if klass doesn't declare its imports
-        # we prefer to set them empty
-        # to avoid clutering
-        klass.imports=[]
-    return klass
 
 
 class WebServerSession:  # ASGI Middleware, for starlette
@@ -154,16 +110,11 @@ class WebServerSession:  # ASGI Middleware, for starlette
         await self.app(scope, receive, send_wrapper)
 
 
-def fqn2hr(fqn:str,js:str,init,session,fullerror=False): # fqn is a "full qualified name", full !
+def normalize(fqn):
     if ":" not in fqn:
         # replace last "." by ":"
         fqn="".join( reversed("".join(reversed(fqn)).replace(".",":",1)))
-
-    klass=getClass(fqn)
-
-    styles=Tag.style("body.htagoff * {cursor:not-allowed !important;}")
-
-    return HRenderer( klass, js, init=init, session = session, fullerror=fullerror, statics=[styles,])
+    return fqn
 
 class HRSocket(WebSocketEndpoint):
     encoding = "text"
@@ -180,62 +131,42 @@ class HRSocket(WebSocketEndpoint):
             return False
 
     async def on_connect(self, websocket):
-        fqn=websocket.path_params.get("fqn","")
-
         await websocket.accept()
-        js="""
-// rewrite the onmessage of the _WS_ to interpret json action now !
-_WS_.onmessage = async function(e) {
-    let actions = await _read_(e.data)
-    action(actions)
-}
-
-// declare the interact js method to communicate thru the WS
-async function interact( o ) {
-    _WS_.send( await _write_(JSON.stringify(o)) );
-}
-
-console.log("started")
-"""
-
-        try:
-            hr=fqn2hr(fqn,js,commons.url2ak(str(websocket.url)),websocket.session,fullerror=websocket.app.debug)
-        except Exception as e:
-            await self._sendback( websocket, str(e) )
-            await websocket.close()
-            return
-
-        self.hr=hr
-
-        # send back the full rendering (1st onmessage after js connection)
-        await self._sendback( websocket, str(self.hr) )
-
-        # register the hr.sendactions, for tag.update feature
-        self.hr.sendactions=lambda actions: self._sendback(websocket,json.dumps(actions))
 
     async def on_receive(self, websocket, data):
+        fqn=websocket.path_params.get("fqn","")
+        uid=websocket.scope["uid"]
+
         if websocket.app.parano:
             data = crypto.decrypt(data.encode(),websocket.app.parano).decode()
-
         data=json.loads(data)
 
-        #=================================== for UT only
-        if data["id"]=="ut":
-            data["id"]=id(self.hr.tag)
-        #===================================
+        p=HrPilot(uid,fqn)
 
-        actions = await self.hr.interact(data["id"],data["method"],data["args"],data["kargs"],data.get("event"))
+        actions=await p.interact( oid=data["id"], method_name=data["method"], args=data["args"], kargs=data["kargs"], event=data.get("event") )
+
         await self._sendback( websocket, json.dumps(actions) )
 
     async def on_disconnect(self, websocket, close_code):
-        del self.hr
+        pass
 
-class AppServer(Starlette):
+def processHrServer():
+    asyncio.run( hrserver() )
+
+
+class AppServer(Starlette):   #NOT THE DEFINITIVE NAME !!!!!!!!!!!!!!!!
     def __init__(self,obj:"htag.Tag class|fqn|None"=None, debug:bool=True,ssl:bool=False,parano:bool=False,sesprovider:"htagweb.sessions.create*|None"=None):
         self.ssl=ssl
+        #TODO: NOT GOOD (when socket disconnet panaro uid change ... so broken !!!!)
         self.parano = str(uuid.uuid4()) if parano else None
         if sesprovider is None:
             sesprovider = sessions.createFile
+        ###################################################################
+
+        p=multiprocessing.Process(target=processHrServer)
+        p.start()
+
+        #################################################################
         Starlette.__init__( self,
             debug=debug,
             routes=[WebSocketRoute("/_/{fqn}", HRSocket)],
@@ -248,8 +179,12 @@ class AppServer(Starlette):
             self.add_route( '/', handleHome )
 
     async def serve(self, request, obj ) -> HTMLResponse:
-        fqn=findfqn(obj)
+        uid = request.scope["uid"]
+        fqn=normalize(findfqn(obj))
+        print(uid,fqn, "->", importClassFromFqn(fqn) )
+
         protocol = "wss" if self.ssl else "ws"
+
         if self.parano:
             jsparano = crypto.JSCRYPTO
             jsparano += f"\nvar _PARANO_='{self.parano}'\n"
@@ -259,77 +194,55 @@ class AppServer(Starlette):
             jsparano = ""
             jsparano += "\nasync function _read_(x) {return x}\n"
             jsparano += "\nasync function _write_(x) {return x}\n"
-        #TODO: consider https://developer.chrome.com/blog/removing-document-write/
 
-        jsbootstrap="""
-            %(jsparano)s
-            // instanciate the WEBSOCKET
-            let _WS_=null;
-            let retryms=500;
-            
-            function connect() {
-                _WS_= new WebSocket("%(protocol)s://"+location.host+"/_/%(fqn)s"+location.search);
-                _WS_.onopen=function(evt) {
-                    console.log("** WS connected")
-                    document.body.classList.remove("htagoff");
-                    retryms=500;
-                    
-                    _WS_.onmessage = async function(e) {
-                        // when connected -> the full HTML page is returned, installed & start'ed !!!
 
-                        let html = await _read_(e.data);
-                        html = html.replace("<body ","<body onload='start()' ");
+        js = """
+%(jsparano)s
 
-                        document.open();
-                        document.write(html);
-                        document.close();
-                    };
-                }
-                
-                _WS_.onclose = function(evt) {
-                    console.log("** WS disconnected");
-                    //console.log("** WS disconnected, retry in (ms):",retryms);
-                    document.body.classList.add("htagoff");
+async function interact( o ) {
+    _WS_.send( await _write_(JSON.stringify(o)) );
+}
 
-                    //setTimeout( function() {
-                    //    connect();
-                    //    retryms=retryms*2;
-                    //}, retryms);
-                };
-            }
+// instanciate the WEBSOCKET
+let _WS_=null;
+let retryms=500;
+
+function connect() {
+    _WS_= new WebSocket("%(protocol)s://"+location.host+"/_/%(fqn)s");
+    _WS_.onopen=function(evt) {
+        console.log("** WS connected")
+        document.body.classList.remove("htagoff");
+        retryms=500;
+        start();
+
+        _WS_.onmessage = async function(e) {
+            let actions = await _read_(e.data)
+            action(actions)
+        };
+
+    }
+
+    _WS_.onclose = function(evt) {
+        console.log("** WS disconnected, retry in (ms):",retryms);
+        document.body.classList.add("htagoff");
+
+        setTimeout( function() {
             connect();
-        """ % locals()
+            retryms=retryms*2;
+        }, retryms);
+    };
+}
+connect();
 
-        # here return a first rendering (only for SEO)
-        # the hrenderer is DESTROYED just after
-        hr=fqn2hr(fqn,jsbootstrap,commons.url2ak(str(request.url)),request.session, fullerror=self.debug)
+""" % locals()
 
-        return HTMLResponse( str(hr) )
+        p = HrPilot(uid,fqn,js)
 
-        # bootstrapHtmlPage="""<!DOCTYPE html>
-        #     <html>
-        #       <head>
-        #             <script>
-        #             %(jsparano)s
-        #             // instanciate the WEBSOCKET
-        #             var _WS_ = new WebSocket("%(protocol)s://"+location.host+"/_/%(fqn)s"+location.search);
-        #             _WS_.onmessage = async function(e) {
-        #                 // when connected -> the full HTML page is returned, installed & start'ed !!!
+        args,kargs = commons.url2ak(str(request.url))
+        html=await p.start(*args,**kargs)
+        return HTMLResponse(html or "no?!")
 
-        #                 let html = await _read_(e.data);
-        #                 html = html.replace("<body ","<body onload='start()' ");
 
-        #                 document.open();
-        #                 document.write(html);
-        #                 document.close();
-        #             };
-        #             </script>
-        #       </head>
-        #       <body>loading</body>
-        #     </html>
-        #     """ % locals()
-
-        # return HTMLResponse( bootstrapHtmlPage )
 
     def run(self, host="0.0.0.0", port=8000, openBrowser=False):   # localhost, by default !!
         if openBrowser:
